requirements.txt
asgiref==3.5.1
django==4.0.4
mysqlclient==2.1.0
django-filter==21.1
django-tables2==2.4.1
django-crispy-forms==1.14.0
django-ckeditor==6.4.0
djangoajax==3.3
django_ajax==0.3.0
django-ajax-tables==1.1.1
dj-database-url
django-import-export
djangorestframework==3.13.1
Markdown==3.3.6
reportlab==3.6.9
pillow==9.1.0
autopep8
gunicorn==20.1.0
pycodestyle==2.8.0
python-decouple==3.6
pytz==2022.1
sqlparse==0.4.2
toml==0.10.2
Unipath==1.1
whitenoise==6.0.0
django-chartjs
xhtml2pdf==0.2.7
django-dynamic-formsets==0.0.8
setuptools==62.1.0
wheel==0.37.1
django-js-asset==2.0.0
